requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
exceptiongroup~=1.2.0
executing==2.2.0
ipykernel==6.29.5
ipython~=8.20.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
Pygments~=2.17.2
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.1
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13

h11~=0.14.0
jupyter~=1.0.0
pip~=24.3.1
attrs~=23.2.0
distro~=1.9.0
wheel~=0.44.0
protobuf~=4.25.3
cryptography~=42.0.7
pillow~=10.2.0
filelock~=3.13.1
future~=1.0.0
requests~=2.31.0
rich~=13.7.1
keras~=3.2.1
ipywidgets~=8.1.1
notebook~=7.0.6
sniffio~=1.3.0
click~=8.1.7
httpcore~=1.0.6
idna~=3.6
httpx~=0.27.2
jiter~=0.7.0
python-dotenv~=1.0.1
openai~=1.58.1
pydantic~=2.9.2
anyio~=4.2.0
numpy~=1.26.4
pandas~=2.2.1
pydantic_core~=2.23.4
certifi~=2023.11.17
annotated-types~=0.7.0
setuptools~=75.3.0